Initial stab to factor out docContext hook
