+++ frontend/src/utilities/index.tsx
@@ -1,3 +1,5 @@
+import { useEffect, useState } from 'react';
+
 /**
  * Converts a Word paragraph object into a usable string by removing leading and trailing
  * spaces and replacing the special Unicode character (that may represent comment).
@@ -18,4 +20,46 @@ export function getParagraphText(paragraphTextObject: Word.Paragraph): string {
 export function handleAutoResize(textarea: HTMLTextAreaElement): void {
 	textarea.style.height = '100%';
 	textarea.style.height = `${textarea.scrollHeight}px`;
+}
+
+/**
+ * Hook to manage the document context in the editor.
+ */
+export function useDocContext(editorAPI: EditorAPI) {
+	const {
+		addSelectionChangeHandler,
+		removeSelectionChangeHandler,
+		getDocContext
+	} = editorAPI;
+
+	const [docContext, updateDocContext] = useState<DocContext>({
+		beforeCursor: '',
+		selectedText: '',
+		afterCursor: ''
+	});
+
+	async function handleSelectionChanged(): Promise<void> {
+		// Get the document context (before cursor, selected text, after cursor)
+		const docInfo = await getDocContext();
+		updateDocContext(docInfo);
+	}
+
+	/**
+	 * useEffect to ensure that event handlers are set up only once
+	 * and cleaned up when the component is unmounted.
+	 * Note that dependencies are empty, so this effect only runs once.
+	 */
+	useEffect(() => {
+		// Handle initial selection change
+		handleSelectionChanged();
+
+		// Handle subsequent selection changes
+		addSelectionChangeHandler(handleSelectionChanged);
+
+		// Cleanup
+		return () => {
+			removeSelectionChangeHandler(handleSelectionChanged);
+		};
+	}, []);
+	return docContext;
 }
